Refactor scheduler api module (#89)
In the future, we may want to call the scheduler pipeline through AMQP
directly, instead of the external scheduler call. This means the
messages sent to the API (via AMQP) may also change. However, currently
the scheduler http api is entangled with the pipeline. This pull request
refactors the api module to support calling the pipeline outside of the
http api.