Add end-to-end tests (#94)
This pull request adds a small pod to the Cortex helm chart, which can
be used to perform commands, such as end-to-end tests. This pod will
have access to the database and also the Cortex configuration, so that
it can make OpenStack calls, making the implementation of e2e tests
significantly easier. In our deployment pipeline, we can use the new pod
to run `kubectl exec -it cortex-cli -- /usr/bin/cortex checks` to
execute the tests. This tooling was also added to the Tilt setup, with a
new command "Run E2E Tests".