Fixed term cortex being redacted in tilt logs (#142)
Replaces the use of `cortex` as a secret in the local config.
Tilt was redacting the term cortex in logs (e.g., `[redacted secret
xxx-password] xxx-password)`, which made things a bit confusing.
With the new secret value, logging should be clearer. No functional
changes—just a small cosmetic/logging tweak.

---------

Co-authored-by: Philipp Matthes <p.matthes@sap.com>