Support multiple pipelines (#262)
Currently cortex only has one pipeline containing scheduler steps that
will be executed for all requests.
However, there are already different kinds of requests which need to be
handled differently.
For example, there are commitment/failover scheduling requests which
need to run filters, while we don't want the filters to run in the
production pipeline (yet). Currently this is done with an obscure flag
sandboxed which should be removed.

With this PR we introduce support for multiple pipelines, which can be
selected in the external scheduler request. We also move the
`preselectAllHosts` flag into the pipeline configuration.

---------

Co-authored-by: Markus Wieland <markus.wieland@sap.com>