Implement commitment-reservation exchange (#296)
This implements the handover between limes + nova and cortex so that vms
are slotted and used commitments no longer reserved.
- The reservations operator now looks for existing virtual machines of
projects that have a commitment. It checks if this commitment matches
the virtual machine(s) and subtract the current usage from the space we
still need to reserve.
- Furthermore, cortex sees incoming initial placement requests matching
a reservation and unlocks it's space for the project.
- Additional refactoring so the reservations operator commitments syncer
handles the filtering of commitments to convert to a reservation