Add commitment change api for limes nova liquid (#299)
Limes can now delegate commitment acceptance to an external service in
its liquids. This implements a new endpoint that takes a commitment
change request, checks it against the initial placement pipeline (if
there is a valid host) and rejects it if not.

These changes will be continued as there are some remaining caveats.