Adding scheduling decision CRD: "why/why not" api  (#307)
This contribution adds an api to cortex that describes why a virtual
machine has been placed on a specific compute host. It exposes the raw
pipeline weights with a kubernetes custom resource, and the associated
operator calculates a human-readable description from the provided
values. The resource is also connected to the nova visualizer so that
decisions can be spectated in the browser removing this dependency on
the mqtt broker.

---------

Co-authored-by: Markus Wieland <44964229+SoWieMarkus@users.noreply.github.com>
Co-authored-by: mblos <156897072+mblos@users.noreply.github.com>