Fixed commitments amount uint64 overflow on more active servers with flavor than committed (#376)
The reservations operator ran into an uint64 overflow when there are
more active servers with the given flavor of a commitment, than
requested by the commitment (`commitment.Amount`).

This was because after iterating the active servers we expected
`commitment.Amount` to be negative if this case occurs. Since
`commitment.Amount` is unsigned, it will never be negative.

In this PR i fixed it by counting the matching servers and checking them
before subtracting them form `commitment.Amount`.

(A cleaner solution might be to not modify the results of limes and
instead create a commitment DTO)