Read + write knowledge in kubernetes only (#415)
Before this change we wrote extracted knowledge into the database.
However, we want to decouple everything but the datasources (raw data)
and feature extraction from the database. This change removes that
knowledges are stored in the database. It also removes the legacy
feature that kpis and scheduler filters could read knowledges from the
database. Accordingly, the kpis and filters that depended on knowledge
tables have been rewritten to be based on the CRs.