KVM infrastructure dashboard (#404)
With the scheduled features for cortex that will only be available in
kvm it makes sense to create a separate kvm infrastructure dashboard.

- track unused reserved capacity per host (cortex reservation CRD)
- track failover capacity (cortex failover CRD?)
- track utilized capacity (hypervisor CRD)
- track total capacity (hypervisor CRD)

For the failover and reserved capacity I am currently exporting dummy
values to be able to already use them inside the dashboard. As soon as
we implemented the features I will use the corresponding CRD as
datasource.
Additionally I will use the hypervisor CRD as datasource.